Add some docs and move classfile manager to independent file
We move the logic of the class file manager to its own file to make it
easier to read and reason about inputs/outputs. This action also
minimizes the length of `Compiler`.

This commit also adds some more docs to the methods that were added
previously and creates `BloopClassFileManager` instead of having a trait
expressing the capability to obtain a set of invalidated class files.
